requirements.txt
playwright
openai
python-dotenv
supabase
